backend/requirements.txt
# FastAPI local development server dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-settings==2.3.3
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
